==> 01_scrapers/03_stats-nba_scrape.ipynb <==
import pandas as pd
import requests
import time
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.common import exceptions

# Scraping Season 2019-2020 stats from stats.nba.com

## Player Traditional Stats

nba_players = 'https://stats.nba.com/players/traditional/?sort=PTS&dir=-1'
driver = webdriver.Chrome(executable_path="../chromedriver/windows/chromedriver.exe")
driver.get(nba_players)
player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
player_stats = []
new_split = []

i = 0
while i < 10:
    try: 
        player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
        
        for q in player:
            player_stats.append(q.text.split('\n'))
            for z in player_stats:
                new_split.append(z[2].split(' '))
                
                    
        button = driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/a[2]')
        button.click()
        i += 1
        
    except exceptions.StaleElementReferenceException:
         pass

player_stats = pd.DataFrame(player_stats).reset_index(drop = True)
new_split = pd.DataFrame(new_split).drop_duplicates(keep = 'first').reset_index(drop = True)
final = pd.concat([player_stats, new_split], axis = 1).drop(columns = [0,2], axis = 1)

trad_columns = ['Player', 'Age', 'W', 'L', 'MP', 'PTS',
               'FGM', 'FGA', 'FG%', '3PM', '3PA', '3P%', 'FTM',
               'FTA', 'FT%', 'OREB', 'DREB', 'REB', 'AST', 'TOV',
               'STL', 'BLK', 'PF', 'FP', 'DD2', 'TD3', '+/-']
final.columns = trad_columns

final.drop(columns = ['PTS', 'STL', 'BLK'], inplace = True)

final.head()

## Player Defense Stats

nba_players = 'https://stats.nba.com/players/defense/?sort=DEF_WS&dir=-1'
driver = webdriver.Chrome(executable_path="../chromedriver/windows/chromedriver.exe")
driver.get(nba_players)
player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
player_stats = []
new_split = []
i = 0
while i < 10:
    try: 
        player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
        
        for q in player:
            player_stats.append(q.text.split('\n'))
            for z in player_stats:
                new_split.append(z[2].split(' '))
                    
        button = driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/a[2]')
        button.click()
        i += 1
        time.sleep(10)
        
    except exceptions.StaleElementReferenceException:
         pass

player_stats = pd.DataFrame(player_stats).reset_index(drop = True)
new_split = pd.DataFrame(new_split).drop_duplicates(keep = 'first').reset_index(drop = True)
defense = pd.concat([player_stats, new_split], axis = 1).drop(columns = [0, 2], axis = 1)

def_columns= ['Player', 'Age', 'GP', 'W', 'L', 'MP', 'DEF_RTG', 'DREB',
            'DREB%', 'T_DREB%', 'STL', 'T_STL%', 'BLK', 'T_BLK%',
            'OP_TOV', 'OP_2NDPTS', 'OP_PAINT', 'DEF_WS']
defense.columns = def_columns
defense.drop(columns = ['Age', 'GP', 'W', 'L', 'MP'], inplace = True)

defense.head()

## Player Usage

nba_players = 'https://stats.nba.com/players/usage/?sort=USG_PCT&dir=-1'
driver = webdriver.Chrome(executable_path="../chromedriver/windows/chromedriver.exe")
driver.get(nba_players)
player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
player_stats = []
new_split = []
i = 0
while i < 10:
    try: 
        player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
        
        for q in player:
            player_stats.append(q.text.split('\n'))
            for z in player_stats:
                new_split.append(z[2].split(' '))
                    
        button = driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/a[2]')
        button.click()
        i += 1
        time.sleep(10)
        
    except exceptions.StaleElementReferenceException:
         pass

player_stats = pd.DataFrame(player_stats).reset_index(drop = True)
new_split = pd.DataFrame(new_split).drop_duplicates(keep = 'first').reset_index(drop = True)
usage = pd.concat([player_stats, new_split], axis = 1).drop(columns = [0, 2], axis = 1)

use_columns = ['Player', 'Age', 'W', 'L', 'MP', 'USG%',
              'T_FGM%', 'T_FGA%', 'T_3PM%', 'T_3PA%', 'T_FTM%',
              'T_FTA%', 'T_OREB%', 'T_DREB%', 'T_REB%', 'T_AST%', 'T_TOV%',
              'T_STL%', 'T_BLK%', 'T_BLKA%', 'T_PF%', 'T_PFD%', 'T_PTS%']
usage.columns = use_columns
usage.drop(columns = ['Age', 'W', 'L', 'MP'], inplace = True)

usage.head()

## Player Shooting Stats

nba_players = 'https://stats.nba.com/players/scoring/?sort=GP&dir=-1'
driver = webdriver.Chrome(executable_path="../chromedriver/windows/chromedriver.exe")
driver.get(nba_players)
player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
player_stats = []
new_split = []
i = 0
while i < 10:
    try: 
        player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
        
        for q in player:
            player_stats.append(q.text.split('\n'))
            for z in player_stats:
                new_split.append(z[2].split(' '))
                    
        button = driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/a[2]')
        button.click()
        i += 1
        time.sleep(10)
        
    except exceptions.StaleElementReferenceException:
         pass

player_stats = pd.DataFrame(player_stats).reset_index(drop = True)
new_split = pd.DataFrame(new_split).drop_duplicates(keep = 'first').reset_index(drop = True)
shooting = pd.concat([player_stats, new_split], axis = 1).drop(columns = [0, 2], axis = 1)

shoot_col = ['Player', 'Age', 'W', 'L', 'MP', 'FGA_2P%', 'FGA_3P%',
            '2PT%', '2PT_MR%', '3PT%', 'FBPTS%', 'FT%', 'OFFTOV%', 'PAINTPTS%',
            'AST_2FGM%', 'UAST_2FGM%', 'AST_3FGM%', 'UAST_3FGM%',
            'FGM_AST%', 'FGM_UAST%']
shooting.columns = shoot_col
shooting.drop(columns = ['W', 'L', 'MP'], inplace= True)

shooting.head()

## Player Misc Stats

nba_players = 'https://stats.nba.com/players/misc/?sort=GP&dir=-1'
driver = webdriver.Chrome(executable_path="../chromedriver/windows/chromedriver.exe")
driver.get(nba_players)
player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
player_stats = []
new_split = []
i = 0
while i < 10:
    try: 
        player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
        
        for q in player:
            player_stats.append(q.text.split('\n'))
            for z in player_stats:
                new_split.append(z[2].split(' '))
                    
        button = driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/a[2]')
        button.click()
        i += 1
        time.sleep(10)
        
    except exceptions.StaleElementReferenceException:
         pass

player_stats = pd.DataFrame(player_stats).reset_index(drop = True)
new_split = pd.DataFrame(new_split).drop_duplicates(keep = 'first').reset_index(drop = True)
misc = pd.concat([player_stats, new_split], axis = 1).drop(columns = [0, 2], axis = 1)

misc_col = ['Player', 'Age', 'W', 'L', 'MP', 'PTS_TOV', '2ND_CH_PTS',
           'FB_PTS', 'PAINT_PTS', 'OPP_PTS_TOV', 'OPP_2ND_PTS', 'OPP_FBPTS',
           'OPP_PAINT_PTS', 'BLK', 'BLKA', 'PF', 'PFD']
misc.columns = misc_col
misc.drop(columns = ['Age', 'W', 'L', 'MP', 'BLK', 'PF'], inplace = True)

# Combining all stats into one dataframe (except opponent)

nba_stats = final.merge(defense, how = 'outer', on = 'Player').merge(usage, how = 'outer', on = 'Player').merge(shooting, how = 'outer', on = 'Player').merge(misc, on = 'Player').drop_duplicates(keep = 'first')

nba_stats.shape

nba_stats.head()

#reworking dataframe in jan 2020
nba_stats.to_csv('../data/nba_later.csv')

-----------------------------------

**The code below this section was only used during my capstone project that was presented on 12/05/2019. Since then, I have been working on coming up with a way to optimize my scrapers So that I would not have to manually clean them**

# adding in salaries and 538 numbers

Adding the salaries dataframe scraped from basketball-reference and the 538 player raptor projects.

salaries = pd.read_csv('../data/salaries.csv')
scrape538 = pd.read_csv('../data/scrape538.csv')
scrape538 = scrape538.drop(['Unnamed: 0', 'Position(s)'], axis = 1)
nba_stats = nba_stats.merge(scrape538, how = 'outer', on = 'Player').merge(salaries, how = 'outer', on = 'Player')

After merging the dataframes, we realize that there are a lot of players with null values. We take care of this by making another dataframe with all of the null values and manually cleaning it. We have null values because each of these sources spell the player names differently. Also, some players were accounted for in the salaries dataframe, however they were recently waived by a their team and are no longer on the roster. Examples of this situation is provided in the technical report. Since I am planning to manually clean the dataframe with all of the null values and later improt the cleaned version, I can drop call null values in the main nba_stats dataframe

problem19 = nba_stats[nba_stats.isna().any(axis=1)]
problem19.to_csv('../data/problem19.csv')
nba_stats.dropna(inplace = True)

problem19.shape

Once I manually cleaned the data, I import the csv back in and concat it to the bottom of the nba_stats dataframe. I also fill in all of the remaining null values. Some null values would be a result of missing salary numbers or 538 rankings. I am not planning on using the salaries as a feature in my model, and will assume if there is a 0, that means the player salary is unknown. If a player doesn't have any 538 rankings, this means that the player has not played enough minutes in a game to recieve a score from 538. Players who have not played that many minutes are irrelevant to my recommendation system because their stats will be inflated anyway (ex. A player who has made all of their FG; however, they have only made one FGA and have played a total of 5 minutes on the court this season).

fixed19 = pd.read_csv('../data/problem19_fixed.csv')
nba_stats = pd.concat([fixed19, nba_stats], ignore_index = True, sort= False).drop(columns = ['Unnamed: 0']).fillna(0)
nba_stats.shape

nba_stats.head()

I export the updated dataframe so that I can use it during EDA and modeling.

--------------------------

string_1 = 'ajgbb wejnfwe'
string_1[0:2:4]

string_1 = 'ajgbb wejnfwe'
string_2 = 'aefnewonrfoawengh'

def checking(str_1, str_2):
    str_1_len = 0
    str_1_list = []
    for i in range(len(str_1)):
        for z in str_1[i]:
            single_str1 = z
            str_1_list.append(single_str1)
    for i2
