==> 01_scrapers/04_bball-ref-scrape.ipynb <==
Scraping based on 100 possessions. 

import pandas as pd
import requests
import time
from time import sleep
from bs4 import BeautifulSoup

url = 'https://www.basketball-reference.com/leagues/NBA_2019_per_poss.html'
res = requests.get(url)

soup = BeautifulSoup(res.content, 'lxml')
table = soup.find('table', {'id': 'per_poss_stats'})

numeric_cols = ['Age', 'G', 'GS', 'MP', 'FG', 'FGA', 'FG%', '3P',
       '3PA', '3P%', '2P', '2PA', '2P%', 'FT', 'FTA', 'FT%', 'ORB', 'DRB',
       'TRB', 'AST', 'STL', 'BLK', 'TOV', 'PF', 'PTS', 'ORtg', 'DRtg']

# BasketballReference Players by Season

# Season '19-20 DataFrame

def grab_stats(year):
    res = requests.get(year)
    soup = BeautifulSoup(res.content, 'lxml')
    table = soup.find('table', {'id' : 'per_poss_stats'})
    statistics = []
    url_list = []
    link = 'https://www.basketball-reference.com'
    
    #creating list of column names
    for label in table.find_all('thead'):
        header = {}
        header = label.findAll('th')
        labels = [i.text for i in header][1:]
    
    #appending row information
    for row in table.find_all('tr')[1:]:
        numbers = [stat.text for stat in row.find_all('td')]
        statistics.append(numbers)
        time.sleep(3)
#         print('this is a row being gathered'
    
    #scrape urls of players
    for z in range(len(table.find_all('tr')[1:])):
        try:
            player_page = link + table.find_all('tr')[1:][z].find('a')['href']
            url_list.append(player_page)
        except:
            error = (f'Error on row {z}')
            url_list.append(error)
            
    df = pd.DataFrame(statistics, columns = labels)
    df['url_list'] = url_list
    df.drop_duplicates(subset = 'Player', keep = 'first').drop(columns = '')
    df.to_csv('data/season1920.csv')
        
    return df
        
season1920 = grab_stats('https://www.basketball-reference.com/leagues/NBA_2020_per_poss.html')

## Season 19-20 Advanced Stats

def grab_stats(year):
    res = requests.get(year)
    soup = BeautifulSoup(res.content, 'lxml')
    table = soup.find('table', {'id' : 'advanced_stats'})
    statistics = []
    
    #creating list of column names
    for label in table.find_all('thead'):
        header = {}
        header = label.findAll('th')
        labels = [i.text for i in header][1:]
    
    #appending row informationv
    for row in table.find_all('tr')[1:]:
        numbers = [stat.text for stat in row.find_all('td')]
        statistics.append(numbers)
        time.sleep(3)
#         print('this is a row being gathered')
            
    df = pd.DataFrame(statistics, columns = labels)
    df.drop_duplicates(subset = 'Player', keep = 'first')
    df.to_csv('data/ad_1920.csv')
        
    return df
        
ad_1920 = grab_stats('https://www.basketball-reference.com/leagues/NBA_2020_advanced.html')
