==> 01_scrapers/05_stats-nba_playerprofiles.ipynb <==
import pandas as pd
import requests
import time
from bs4 import BeautifulSoup
from selenium import webdriver
from selenium.common import exceptions

# Player Bios from stats.nba

player_bio[1][1]

nba_players = 'https://stats.nba.com/players/bio/?sort=PLAYER_WEIGHT&dir=1'
driver = webdriver.Chrome(executable_path="../chromedriver/windows/chromedriver.exe")
driver.get(nba_players)
player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
player_bio = []
new_split = []

i = 0
while i < 10:
    try: 
        player = driver.find_elements_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr')
        
        for q in player:
            player_bio.append(q.text.split('\n'))
            for z in player_bio:
                new_split.append(z[1].split(' '))
                
                    
        button = driver.find_element_by_xpath('/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[1]/div/div/a[2]/i')
        button.click()
        i += 1
        
    except exceptions.StaleElementReferenceException:
         pass

player_bio = pd.DataFrame(player_bio).reset_index(drop = True)
new_split = pd.DataFrame(new_split).drop_duplicates(keep = 'first').reset_index(drop = True)
player_bio = pd.concat([player_bio, new_split], axis = 1)

columns

url = 'https://stats.nba.com/players/bio/?sort=PLAYER_NAME&dir=-1'
driver = webdriver.Chrome(executable_path = '../chromedriver/windows/chromedriver.exe')
driver.get(url)

html = driver.page_source
html = BeautifulSoup(html, 'lxml')
col_names = html.find('table').find_all('th')
column = [i.text for i in col_names]

column

Player images

nba_players = 'https://stats.nba.com/players/bio/?sort=PLAYER_WEIGHT&dir=1'
driver = webdriver.Chrome(executable_path="../chromedriver/windows/chromedriver.exe")
driver.get(nba_players)

html = driver.page_source
html = BeautifulSoup(html, 'lxml')
table = html.find('table')

team_link = []
for i in table.find_all('tr'):
    player_name = i.find('td')
    player_name.find('a')
#     player_profiles = {}
#     player_profiles['Player'] = i.text
#     player_profiles['href'] = i.find('a').attrs['href']
#     team_link.append(player_profiles)

profile = driver.find

tbody = table.find('tbody')
player = tbody.find_all('tr')
player

/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr[1]/td[1]/a
/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[2]/table/tbody/tr[2]/td/a
/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[2]/table/tbody/tr[3]


/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[1]/table/tbody/tr[1]/td[1]/a
/html/body/main/div[2]/div/div[2]/div/div/nba-stat-table/div[2]/div[2]/table/tbody/tr[5]/td/a

Plan A that isn't working - grab the hrefs of each profile and then enter profile by entering with Beautiful Soup

Plan b - take the xpath of each of the player profiles i+1 on the tr[] expat

Grab image url from player profile

url = 'https://stats.nba.com/player/1626246/'
res = requests.get(url)
soup = BeautifulSoup(res.content, 'lxml')
playerprofile = soup.find('div', {'class' : 'player-summary__image-block'})
playerprofile.find('img')
